[FIX] website_mass_mailing loading
