+++ addons/website_mass_mailing/views/website_mass_mailing_templates.xml
@@ -12,6 +12,7 @@
 
 <template id="assets_editor" inherit_id="website.assets_wysiwyg">
     <xpath expr="//script[last()]" position="after">
+        <script type="text/javascript" src="/website_mass_mailing/static/src/js/wysiwyg.js"/>
         <script type="text/javascript" src="/website_mass_mailing/static/src/js/website_mass_mailing.editor.js"/>
     </xpath>
 </template>
